Fixed ExampleCreator

+++ Assets/Scripts/ML-Agents/SaveGridToFileScript.cs
@@ -10,6 +10,18 @@ public class SerializedGridObject
     public Vector3 rotation;
 }
 
+[System.Serializable]
+public class SerializedGrid
+{
+    public SerializedGridObject[] gridObjects;
+}
+
+[System.Serializable]
+public class SerializedGrids
+{
+    public SerializedGrid[] grids;
+}
+
 public class SaveGridToFileScript : MonoBehaviour
 {
     [SerializeField] private ObjectPlacer objectPlacer;
@@ -17,47 +29,66 @@ public class SaveGridToFileScript : MonoBehaviour
     public void OnSave()
     {
         var placedGameObjects = objectPlacer.GetPlacedGameObjects();
-        List<SerializedGridObject> serializedGridObjects = new List<SerializedGridObject>();
-        
+        var serializedGrid = new SerializedGrid();
+        serializedGrid.gridObjects = new SerializedGridObject[placedGameObjects.Count];
+
         foreach (var placedGameObject in placedGameObjects)
         {
             var serializedGridObject = new SerializedGridObject();
             serializedGridObject.name = placedGameObject.name.Substring(0, placedGameObject.name.Length - 7);
             serializedGridObject.position = placedGameObject.transform.position;
             serializedGridObject.rotation = placedGameObject.transform.rotation.eulerAngles;
-            serializedGridObjects.Add(serializedGridObject);
+            serializedGrid.gridObjects[placedGameObjects.IndexOf(placedGameObject)] = serializedGridObject;
         }
+
+        var path = Application.dataPath + "/ML-Agents/Grids.json";
         
-        string json = JsonUtility.ToJson(serializedGridObjects);
-        string path = UnityEditor.EditorUtility.SaveFilePanel(
-            "Save grid to json",
-            "",
-            "grid.json",
-            "json");
+        var serializedGrids = new SerializedGrids();
+        serializedGrids.grids = new SerializedGrid[1];
         
-        System.IO.File.WriteAllText(path, json);
-        
-        Debug.Log(json);
-    }
-    
-    public void LoadGrid()
-    {
-        string path = UnityEditor.EditorUtility.OpenFilePanel(
-            "Load grid from json",
-            "",
-            "json");
-        
-        string json = System.IO.File.ReadAllText(path);
-        
-        List<SerializedGridObject> serializedGridObjects = JsonUtility.FromJson<List<SerializedGridObject>>(json);
-        
-        foreach (var serializedGridObject in serializedGridObjects)
+        if(System.IO.File.Exists(path))
         {
-            var prefab = Resources.Load<GameObject>(serializedGridObject.name);
-            objectPlacer.PlaceObject(prefab, Quaternion.Euler(serializedGridObject.rotation.x, serializedGridObject.rotation.y, serializedGridObject.rotation.z), new Vector3(serializedGridObject.position.x, serializedGridObject.position.y, serializedGridObject.position.z));
+            var json = System.IO.File.ReadAllText(path);
+            serializedGrids = JsonUtility.FromJson<SerializedGrids>(json);
+        
+            //add new grid to grids
+            var grids = serializedGrids.grids;
+            var newGrids = new SerializedGrid[grids.Length + 1];
+            for (int i = 0; i < grids.Length; i++)
+            {
+                newGrids[i] = grids[i];
+            }
+            newGrids[grids.Length] = serializedGrid;
+            serializedGrids.grids = newGrids;
+        
+            var json2 = JsonUtility.ToJson(serializedGrids);
+            System.IO.File.WriteAllText(path, json2);
+        } 
+        else
+        {
+            serializedGrids.grids[0] = serializedGrid;
+            var json = JsonUtility.ToJson(serializedGrids);
+            System.IO.File.WriteAllText(path, json);
         }
         
-        Debug.Log(json);
+        Debug.Log($"Saved grid to {path}");
+        
     }
 
+    public SerializedGrids GetGrids()
+    {
+        var path = Application.dataPath + "/ML-Agents/Grids.json";
+        
+        if(!System.IO.File.Exists(path))
+        {
+            Debug.LogError($"File {path} does not exist");
+            return null;
+        }
+        
+        var json = System.IO.File.ReadAllText(path);
+        
+        var serializedGrids = JsonUtility.FromJson<SerializedGrids>(json);
+        
+        return serializedGrids;
+    }
 }
